fix defects and compression and optimize folder structure
