+++ public/robots.txt
@@ -1,2 +1,5 @@
-User-agent: *
-Disallow: /
+User-agent: Googlebot
+allow: /
+
+User-Agent:  FacebookBot
+allow: /
