+++ src/block/useBlock.tsx
@@ -2,57 +2,35 @@ import { useCallback, useContext, useEffect, useRef, useState } from 'react';
 import { isEqual } from 'lodash';
 import { useDrag } from 'react-dnd';
 import { getEmptyImage } from 'react-dnd-html5-backend';
+
 import { macMahon, palettesMap } from '../common/constants';
-import {
-	BlockAnimation,
-	BlockList,
-	BumperColorAndCoordinates,
-	Cursor,
-	Opacity,
-	Pattern,
-	PatternDirection,
-	TransitionDuration,
-	ZIndex,
-} from '../common/interfaces';
+import { BlockAnimation, Display, PatternDirection, TransitionDuration, ZIndex } from '../common/enums';
 import {
 	BumperContext,
 	StagedContext,
 	ThemeContext,
 	ThrottleContext,
 	BenchContext,
-	BlockContextProps,
-	ThrottleContextProps,
-	BumperContextProps,
-	ThemeContextProps,
-	KidsModeContextProps,
 	KidsModeContext,
-	OrientationContextProps,
 	OrientationContext,
-	AnchorLegConextProps,
 	AnchorLegContext,
 } from '../common/contexts';
 import { rotateClockwise, rotateAntiClockwise } from '../common/utils';
-import { BlockProps } from './Block';
-import seatService from '../seat/seatService';
 
-interface UseBlockReturn {
-	id?: { id: string };
-	colors: string[];
-	dragRef?: any;
-	dndStyle?: {
-		opacity: Opacity;
-		zIndex: ZIndex;
-		cursor: Cursor;
-	};
-	rotateStyle?: {
-		transform: string;
-		transitionDuration: TransitionDuration;
-	};
-	backLayerStyle?: {
-		animationName: string;
-		animationDuration: TransitionDuration;
-	};
-}
+import seatService from '../seat/seatService';
+import {
+	AnchorLegConextProps,
+	BlockContextProps,
+	BlockList,
+	BumperContextProps,
+	KidsModeContextProps,
+	OrientationContextProps,
+	Pattern,
+	ThemeContextProps,
+	ThrottleContextProps,
+} from '../common/common.types';
+import { BlockProps, UseBlockReturn } from './Block.types';
+import { BumperColorAndCoordinates } from '../panel/Panel.types';
 
 const useBlock: (props: BlockProps) => UseBlockReturn = ({
 	isStage,
@@ -75,8 +53,10 @@ const useBlock: (props: BlockProps) => UseBlockReturn = ({
 	const [animationName, setAnimationName] = useState<BlockAnimation>(BlockAnimation.EMPTY);
 	const [zIndex, setZIndex] = useState(ZIndex.BACK);
 	const freeze = useRef<boolean>(false);
-	const blockRef = useRef<HTMLDivElement>();
+	const blockDOMRef = useRef<HTMLDivElement>();
+	const blockRef = useRef<Pattern>([0, 0, 0, 0]);
 	const bumperColorRef = useRef<number>();
+	const blockListRef = useRef<BlockList>(blockList);
 	const triggerSingleClick = useRef<boolean>(true);
 
 	const getWidth = useCallback(
@@ -84,13 +64,12 @@ const useBlock: (props: BlockProps) => UseBlockReturn = ({
 		[isLandscape, stageOrientationLock]
 	);
 
-	const [{ opacity, cursor }, dragRef, dragPreview] = useDrag(
+	const [{ display }, dragRef, dragPreview] = useDrag(
 		() => ({
 			type: macMahon,
 			item: { isStage, isLandscape, seatNumber, pattern },
 			collect: monitor => ({
-				opacity: monitor.isDragging() ? Opacity.INVISABLE : Opacity.VISABLE,
-				cursor: monitor.isDragging() ? Cursor.DRAGGING : Cursor.DRAG,
+				display: monitor.isDragging() ? Display.HIDDEN : Display.SHOW,
 			}),
 			canDrag: !freeze.current,
 		}),
@@ -105,10 +84,10 @@ const useBlock: (props: BlockProps) => UseBlockReturn = ({
 				kidsMode,
 				isLandscape,
 				stageOrientationLock,
-				updatedStagedBlockList: blockList,
+				updatedStagedBlockList: blockListRef.current,
 				seatNumber,
 				movedPattern: updatedBlock || [0, 0, 0, 0],
-				bumperColor: bumper.bumperColor,
+				bumperColor: bumperColorRef.current,
 			})
 		);
 	};
@@ -126,17 +105,13 @@ const useBlock: (props: BlockProps) => UseBlockReturn = ({
 						isSourceFromStage: true,
 						isTargetingStage: true,
 						movedPattern: updatedBlock || [0, 0, 0, 0],
-						stagedBlockList: blockList,
+						stagedBlockList: blockListRef.current,
 						updatedStagedBlockList: updatedBlockList,
 						seatNumber,
 					}),
 			  }
 			: undefined;
 	};
-
-	useEffect(() => {
-		bumperColorRef.current = bumper.bumperColor;
-	}, [bumper.bumperColor]);
 
 	const rotateBlock = (deltaY: number): void => {
 		const toClockwise = deltaY > 0;
@@ -146,12 +121,12 @@ const useBlock: (props: BlockProps) => UseBlockReturn = ({
 		setZIndex(ZIndex.FRONT);
 		freeze.current = true;
 
-		const rotatedBlock = blockList[seatNumber];
+		const rotatedBlock = blockRef;
 		if (!rotatedBlock) {
 			return;
 		}
 
-		const updatedBlock = toClockwise ? rotateClockwise(rotatedBlock) : rotateAntiClockwise(rotatedBlock);
+		const updatedBlock = (toClockwise ? rotateClockwise : rotateAntiClockwise)(blockRef.current);
 		if (needsBounceBack(updatedBlock)) {
 			setTimeout(() => {
 				setRotate(0);
@@ -162,11 +137,15 @@ const useBlock: (props: BlockProps) => UseBlockReturn = ({
 			}, 70);
 			return;
 		}
-		const updatedBlockList = blockList.map((block, i) => (i === seatNumber ? updatedBlock : block));
-		const updatedBumper = checkAndUpdateBumper(updatedBlock, updatedBlockList);
 		setTimeout(() => {
-			setBumper({ ...bumper, ...updatedBumper, bumperColor: bumperColorRef.current });
-			setBlockList(updatedBlockList);
+			const prePolitUpdatedBlockList = blockListRef.current.map((block, i) =>
+				i === seatNumber ? updatedBlock : block
+			);
+			const updatedBumper = checkAndUpdateBumper(updatedBlock, prePolitUpdatedBlockList);
+			isStage && setBumper({ ...bumper, ...updatedBumper, bumperColor: bumperColorRef.current });
+			setBlockList(blockList => {
+				return blockList.map((block, i) => (i === seatNumber ? updatedBlock : block));
+			});
 			setTransitionDuration(TransitionDuration.STATIC);
 			setRotate(0);
 			setZIndex(ZIndex.BACK);
@@ -191,31 +170,54 @@ const useBlock: (props: BlockProps) => UseBlockReturn = ({
 		event.preventDefault();
 		triggerSingleClick.current = true;
 		setTimeout(() => {
-			triggerSingleClick.current && rotateBlock(-90);
+			triggerSingleClick.current && rotateBlock(90);
+			blockDOMRef.current?.removeEventListener('click', onSingleTap);
+			blockDOMRef.current?.removeEventListener('dblclick', onDoubleTap);
+			setTimeout(() => {
+				blockDOMRef.current?.addEventListener('click', onSingleTap);
+				blockDOMRef.current?.addEventListener('dblclick', onDoubleTap);
+			}, 310);
 		}, 180);
 	};
 
 	const onDoubleTap = (event: Event) => {
 		event.preventDefault();
 		triggerSingleClick.current = false;
-		rotateBlock(90);
+		blockDOMRef.current?.removeEventListener('click', onSingleTap);
+		blockDOMRef.current?.removeEventListener('dblclick', onDoubleTap);
+		setTimeout(() => {
+			blockDOMRef.current?.addEventListener('click', onSingleTap);
+			blockDOMRef.current?.addEventListener('dblclick', onDoubleTap);
+		}, 310);
+		rotateBlock(-90);
 	};
 
 	const onWheelScroll = (event: WheelEvent) => onThrottleWheelScroll(event.deltaY, rotateBlock);
 
 	useEffect(() => {
-		blockRef.current?.addEventListener('wheel', onWheelScroll, { passive: false });
-		blockRef.current?.addEventListener('click', onSingleTap);
-		blockRef.current?.addEventListener('dblclick', onDoubleTap);
-		onClickOverride && blockRef.current?.addEventListener('click', onClickOverride);
+		const updatedBlock = blockList[seatNumber];
+		if (updatedBlock) {
+			blockRef.current = updatedBlock;
+		}
+		blockListRef.current = blockList;
+	}, [blockList]);
+
+	useEffect(() => {
+		blockDOMRef.current?.addEventListener('wheel', onWheelScroll, { passive: false });
+		if (onClickOverride) {
+			blockDOMRef.current?.addEventListener('click', onClickOverride);
+		} else {
+			blockDOMRef.current?.addEventListener('click', onSingleTap);
+			blockDOMRef.current?.addEventListener('dblclick', onDoubleTap);
+		}
 
 		return () => {
-			blockRef.current?.removeEventListener('wheel', onWheelScroll);
-			blockRef.current?.removeEventListener('click', onSingleTap);
-			blockRef.current?.removeEventListener('dblclick', onDoubleTap);
-			onClickOverride && blockRef.current?.removeEventListener('click', onClickOverride);
+			blockDOMRef.current?.removeEventListener('wheel', onWheelScroll);
+			blockDOMRef.current?.removeEventListener('click', onSingleTap);
+			blockDOMRef.current?.removeEventListener('dblclick', onDoubleTap);
+			onClickOverride && blockDOMRef.current?.removeEventListener('click', onClickOverride);
 		};
-	}, [isStage, seatNumber, palettes, blockList]);
+	}, [isStage, seatNumber, palettes]);
 
 	useEffect(() => {
 		dragPreview(getEmptyImage(), { captureDraggingState: true });
@@ -232,6 +234,10 @@ const useBlock: (props: BlockProps) => UseBlockReturn = ({
 		}
 	}, [anchorLeg]);
 
+	useEffect(() => {
+		bumperColorRef.current = bumper.bumperColor;
+	}, [bumper.bumperColor]);
+
 	const colors = pattern.map(color => (palettes || palettesMap[contextPalettes])[color]);
 	if (seatNumber < 0) {
 		const backgroundImage = `conic-gradient(at center, ${colors[PatternDirection.TOP]} 12.5%, 
@@ -244,7 +250,7 @@ const useBlock: (props: BlockProps) => UseBlockReturn = ({
 			frontLayerStyle: {
 				backgroundImage,
 			},
-			dragRef: dragRef(blockRef),
+			dragRef: blockDOMRef,
 		};
 	}
 	const id =
@@ -262,11 +268,10 @@ const useBlock: (props: BlockProps) => UseBlockReturn = ({
 	return {
 		id,
 		colors,
-		dragRef: dragRef(blockRef),
+		dragRef: dragRef(blockDOMRef),
 		dndStyle: {
-			opacity,
+			display,
 			zIndex,
-			cursor,
 			filter: darkTheme ? 'brightness(0.8)' : '',
 		},
 		rotateStyle: {
