+++ src/common/useLocalStorage.tsx
@@ -1,14 +1,7 @@
 import { useState } from 'react';
-import { StageOrientationLock } from './interfaces';
-import { localStorageSettings } from './constants';
 
-interface SettingsProps {
-	palettes?: string;
-	stageOrientationLock?: StageOrientationLock;
-	kidsMode?: boolean;
-	darkTheme?: boolean;
-	tutorialStep?: number;
-}
+import { SettingsProps } from './common.types';
+import { localStorageSettings } from './constants';
 
 const useLocalStorage = () => {
 	const storedSettings = JSON.parse(window.localStorage.getItem(localStorageSettings) || '{}');
