+++ src/menu/Menu.types.ts
@@ -1,0 +1,3 @@
+import { ReactElement } from 'react';
+
+export type SvgButton = (onclick?: () => void) => ReactElement;
