+++ src/panel/Panel.styles.ts
