+++ src/panel/StagePanel.tsx
@@ -1,11 +1,11 @@
 import { CSSProperties, FC, memo } from 'react';
 import { css } from 'aphrodite/no-important';
 import { ThrottleContext } from '../common/contexts';
-import { StageOrientationLock } from '../common/interfaces';
+import { StageOrientationLock } from '../common/enums';
 import { darkBumperColor, lightBumperColor, palettesMap } from '../common/constants';
 import usePanel from './usePanel';
 import Seat from '../seat/Seat';
-import styles from './Panel.style';
+import styles from './Panel.styles';
 
 const bumperStyle: CSSProperties = {
 	position: 'relative',
